Add loading component, better responsiveness, fix error message not showing if sending non array

+++ src/views/DeleteRecipesView.vue
@@ -1,20 +1,27 @@
 <template>
-  <div v-show="loadingText" class="loading-text">{{ loadingText }}</div>
+  <LoadingText v-show="loadingText" :loading-text="loadingText"></LoadingText>
+
+  <h1>Ta bort recept</h1>
 
   <form @submit.prevent="deleteRecipes()">
-    <input
+    <textarea
       type="string"
       v-model="recipes"
       placeholder="Array med recept (måste tas direkt från APIt)"
-    />
-    <button type="submit" :disabled="fetching">Ta bort recepten</button>
+    ></textarea>
+    <button type="submit" :disabled="fetching || !recipes">Ta bort recepten</button>
   </form>
 </template>
 
 <script>
+import LoadingText from '@/components/LoadingText.vue';
 import { useTeamIdStore } from '@/stores/teamId';
 
 export default {
+  components: {
+    LoadingText,
+  },
+
   setup() {
     const store = useTeamIdStore();
     return { store };
@@ -36,56 +43,59 @@ export default {
 
   methods: {
     deleteRecipes() {
-      this.loadingText = 'Raderar recept...';
-      this.parsedRecipes.forEach(async (recipe) => {
-        try {
-          const response = await fetch(
-            `https://recipes.bocs.se/api/v1/${this.store.teamId}/recipes/${recipe.id}`,
-            {
-              method: 'DELETE',
-              headers: { 'Content-type': 'application/json' },
+      if (this.parsedRecipes) {
+        this.fetching = true;
+        this.loadingText = 'Raderar recepten...';
+        this.parsedRecipes.forEach(async (recipe) => {
+          try {
+            const response = await fetch(
+              `https://recipes.bocs.se/api/v1/${this.store.teamId}/recipes/${recipe.id}`,
+              {
+                method: 'DELETE',
+                headers: { 'Content-type': 'application/json' },
+              }
+            );
+            if (!response.ok) {
+              throw new Error(`Status: ${response.status}`);
             }
-          );
-          if (!response.ok) {
-            throw new Error(`Status: ${response.status}`);
+            this.loadingText = 'Recept borttagna';
+            setTimeout(() => {
+              this.loadingText = '';
+            }, 1000);
+          } catch (error) {
+            this.loadingText = `Fel inträffade vid senaste försöket med ${error.message.toLowerCase()}`;
+            console.error('Fetch failed:', error);
+          } finally {
+            this.fetching = false;
           }
-          this.loadingText = 'Recept borttagna';
-          setTimeout(() => {
-            this.loadingText = '';
-          }, 1000);
-        } catch (error) {
-          console.error('Fetch failed:', error);
-          this.loadingText = `Fel inträffade vid senaste försöket med ${error.message.toLowerCase()}`;
-        } finally {
-          this.fetching = false;
-        }
-      });
+        });
+      } else {
+        this.loadingText = 'Fel inträffade, se till att du skickar en array';
+      }
     },
   },
 };
 </script>
 
 <style scoped>
-.loading-text {
-  position: fixed;
-  right: 5em;
-  top: 5em;
-  color: white;
-  background-color: black;
-}
-
-input {
-  width: 50%;
-  margin: 1em;
-  height: 2em;
-}
-
 form {
   padding: 1em;
+  display: flex;
+  flex-direction: column;
 }
 
 .recipe-forms {
   display: flex;
   flex-direction: column;
+}
+
+textarea {
+  width: 50%;
+  margin: 1em;
+  height: 20em;
+}
+
+button {
+  width: 9em;
 }
 </style>
